changes Working insert for Gabinete

+++ Common/Colores.cs
@@ -7,7 +7,7 @@ namespace Univalle.Fie.Sistemas.BaseDeDatos2.AppComputadorasBDD.Common
 {
     public class Colores
     {
-        public int IdColor { get; set; }
+        public short IdColor { get; set; }
         public string Nombre { get; set; }
         public Color ColorRGB { get; set; }
     }
